backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Computer Vision & ML
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
mmcv==2.1.0
mmdet==3.2.0
mmpose==1.1.0
mediapipe==0.10.7

# Supabase integration
supabase==2.0.2
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0
matplotlib==3.8.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 